# refactor(qrcode): use kmp to speed up rule3 pattern calculating.