use `kmp` algorithm to optimize speed up `rule3` (find sub pattern in matrix) calculating process.