@yeqown 2021-12-21
old benchmark:
```sh
goos: darwin
goarch: amd64
pkg: github.com/yeqown/go-qrcode/v2
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
Benchmark_NewQRCode_1KB-4   	      79	  18715841 ns/op
Benchmark_NewQRCode_1KB-4   	      82	  17220630 ns/op
Benchmark_NewQRCode_1KB-4   	      79	  23198089 ns/op
Benchmark_NewQRCode_1KB-4   	      80	  20673591 ns/op
Benchmark_NewQRCode_1KB-4   	      50	  23134053 ns/op
PASS
ok  	github.com/yeqown/go-qrcode/v2	8.849s
```

new benchmark:
```sh
goos: darwin
goarch: amd64
pkg: github.com/yeqown/go-qrcode/v2
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
Benchmark_NewQRCode_1KB-4   	     100	  12565710 ns/op
Benchmark_NewQRCode_1KB-4   	     100	  19004634 ns/op
Benchmark_NewQRCode_1KB-4   	     138	   9097741 ns/op
Benchmark_NewQRCode_1KB-4   	     100	  11519712 ns/op
Benchmark_NewQRCode_1KB-4   	     130	   8298252 ns/op
Benchmark_NewQRCode_1KB-4   	     141	   7578266 ns/op
Benchmark_NewQRCode_1KB-4   	     150	   8567960 ns/op
Benchmark_NewQRCode_1KB-4   	     144	   8052277 ns/op
Benchmark_NewQRCode_1KB-4   	     144	  11689046 ns/op
Benchmark_NewQRCode_1KB-4   	     122	   9214623 ns/op
PASS
ok  	github.com/yeqown/go-qrcode/v2	20.525s
```

**delta display**:

```sh
benchstat old.txt new.txt
name              old time/op  new time/op  delta
_NewQRCode_1KB-4  20.6ms ±16%   9.6ms ±31%  -53.27%  (p=0.001 n=5+9)

```